+++ crates/re_log_types/src/data.rs
@@ -10,6 +10,7 @@ use self::data_types::Vec3;
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub enum DataType {
     // 1D:
+    Bool,
     I32,
     F32,
     Color,
@@ -60,11 +61,18 @@ pub mod data_types {
         }
     }
 
+    impl DataTrait for bool {
+        fn data_typ() -> DataType {
+            DataType::Bool
+        }
+    }
+
     impl DataTrait for i32 {
         fn data_typ() -> DataType {
             DataType::I32
         }
     }
+
     impl DataTrait for f32 {
         fn data_typ() -> DataType {
             DataType::F32
@@ -155,6 +163,7 @@ pub mod data_types {
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub enum Data {
     // 1D:
+    Bool(bool),
     I32(i32),
     F32(f32),
     Color(data_types::Color),
@@ -191,6 +200,7 @@ impl Data {
     #[inline]
     pub fn data_type(&self) -> DataType {
         match self {
+            Self::Bool(_) => DataType::Bool,
             Self::I32(_) => DataType::I32,
             Self::F32(_) => DataType::F32,
             Self::Color(_) => DataType::Color,
@@ -214,6 +224,7 @@ impl Data {
     }
 }
 
+impl_into_enum!(bool, Data, Bool);
 impl_into_enum!(i32, Data, I32);
 impl_into_enum!(f32, Data, F32);
 impl_into_enum!(BBox2D, Data, BBox2D);
@@ -231,6 +242,7 @@ impl_into_enum!(Transform, Data, Transform);
 #[derive(Clone, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub enum DataVec {
+    Bool(Vec<bool>),
     I32(Vec<i32>),
     F32(Vec<f32>),
     Color(Vec<data_types::Color>),
@@ -265,6 +277,7 @@ pub enum DataVec {
 macro_rules! data_map(
     ($data: expr, |$value: pat_param| $action: expr) => ({
         match $data {
+            $crate::Data::Bool($value) => $action,
             $crate::Data::I32($value) => $action,
             $crate::Data::F32($value) => $action,
             $crate::Data::Color($value) => $action,
@@ -294,6 +307,7 @@ macro_rules! data_map(
 macro_rules! data_vec_map(
     ($data_vec: expr, |$vec: pat_param| $action: expr) => ({
         match $data_vec {
+            $crate::DataVec::Bool($vec) => $action,
             $crate::DataVec::I32($vec) => $action,
             $crate::DataVec::F32($vec) => $action,
             $crate::DataVec::Color($vec) => $action,
@@ -316,6 +330,7 @@ impl DataVec {
     #[inline]
     pub fn element_data_type(&self) -> DataType {
         match self {
+            Self::Bool(_) => DataType::Bool,
             Self::I32(_) => DataType::I32,
             Self::F32(_) => DataType::F32,
             Self::Color(_) => DataType::Color,
@@ -348,6 +363,7 @@ impl DataVec {
 
     pub fn last(&self) -> Option<Data> {
         match self {
+            Self::Bool(vec) => vec.last().cloned().map(Data::Bool),
             Self::I32(vec) => vec.last().cloned().map(Data::I32),
             Self::F32(vec) => vec.last().cloned().map(Data::F32),
             Self::Color(vec) => vec.last().cloned().map(Data::Color),