+++ rerun_py/src/python_bridge.rs
@@ -119,6 +119,7 @@ fn rerun_sdk(py: Python<'_>, m: &PyModule) -> PyResult<()> {
 
     m.add_function(wrap_pyfunction!(log_mesh_file, m)?)?;
     m.add_function(wrap_pyfunction!(log_image_file, m)?)?;
+    m.add_function(wrap_pyfunction!(set_visible, m)?)?;
 
     Ok(())
 }
@@ -1445,6 +1446,24 @@ fn log_image_file(
         &time_point,
         (&obj_path, "space"),
         LoggedData::Single(Data::Space(parse_obj_path(&space)?)),
+    );
+
+    Ok(())
+}
+
+/// Clear the visibility flag of an object
+#[pyfunction]
+fn set_visible(obj_path: &str, visibile: bool) -> PyResult<()> {
+    let obj_path = parse_obj_path(obj_path)?;
+
+    let mut sdk = Sdk::global();
+
+    let time_point = time(false);
+
+    sdk.send_data(
+        &time_point,
+        (&obj_path, "_visible"),
+        LoggedData::Single(Data::Bool(visibile)),
     );
 
     Ok(())